Extract lsystem-app-model crate
Add a new toolkit-free crate that consolidates application-model
logic shared by both GUI crates (Iced and Leptos):

- config_workspace: moved from lsystem-core so that crate stays
  focused on L-system semantics
- presets: single load_presets() replaces identical helpers in both
  GUI crates
- color: unified LineColorMode (ALL, Display, from_key/key) and
  ColorControlMemory, replacing per-crate duplicates
- animation: canonical HsvMovement without phase_degrees field;
  phase stored externally per GUI (StoredValue in Leptos, f32 field
  on FractalApp in Iced); advance_hsv_phase_degrees free function
- util: sanitize_filename, rgb_to_hex, hex_to_rgb consolidated from
  two GUI crates; tests now run on native targets (were wasm-only)

Also:
- Add contains_3d_symbols to lsystem-core/alphabet.rs, re-exported
  from lsystem-core; replaces identical local fns in both GUI crates
- Promote ConfigSource setters to pub so config_workspace can call
  them from the new crate
- Remove four unused web-sys features from lsystem-web-app
- Update AGENTS.md and CONTRIBUTING.md for the new crate
