+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "crates/lsystem-core",
     "crates/lsystem-renderer",
+    "crates/lsystem-app-model",
     "crates/lsystem-app",
     "crates/lsystem-web-app",
 ]
@@ -20,6 +21,7 @@ js-sys = "0.3"
 log = "0.4"
 lsystem-core = { path = "crates/lsystem-core" }
 lsystem-renderer = { path = "crates/lsystem-renderer" }
+lsystem-app-model = { path = "crates/lsystem-app-model" }
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
 web-sys = "0.3"
