+++ crates/lsystem-app-model/src/animation.rs
@@ -1,16 +1,18 @@
 use lsystem_core::LineColorConfig;
 
-pub(crate) const HSV_MOVEMENT_DEFAULT_SPEED_DEGREES_PER_SECOND: f32 = 15.0;
-pub(crate) const HSV_MOVEMENT_MIN_SPEED_DEGREES_PER_SECOND: f32 = 1.0;
-pub(crate) const HSV_MOVEMENT_MAX_SPEED_DEGREES_PER_SECOND: f32 = 60.0;
+pub const HSV_MOVEMENT_DEFAULT_SPEED_DEGREES_PER_SECOND: f32 = 15.0;
+pub const HSV_MOVEMENT_MIN_SPEED_DEGREES_PER_SECOND: f32 = 1.0;
+pub const HSV_MOVEMENT_MAX_SPEED_DEGREES_PER_SECOND: f32 = 60.0;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum HsvMovementDirection {
+pub enum HsvMovementDirection {
     Forward,
     Reverse,
 }
 
 impl HsvMovementDirection {
+    pub const ALL: &'static [Self] = &[Self::Forward, Self::Reverse];
+
     pub(crate) fn sign(self) -> f32 {
         match self {
             Self::Forward => 1.0,
@@ -19,48 +21,58 @@ impl HsvMovementDirection {
     }
 }
 
+impl std::fmt::Display for HsvMovementDirection {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(match self {
+            Self::Forward => "Forward",
+            Self::Reverse => "Reverse",
+        })
+    }
+}
+
 /// User-driven configuration for HSV hue-cycle animation.
-/// The phase accumulator is stored separately as a non-reactive `StoredValue<f32>`.
+/// The phase accumulator is stored separately by each GUI:
+/// Leptos uses `StoredValue<f32>`, Iced uses a plain `f32` field on `FractalApp`.
 #[derive(Clone, Copy, Debug)]
-pub(crate) struct HsvMovement {
+pub struct HsvMovement {
     enabled: bool,
     speed_degrees_per_second: f32,
     direction: HsvMovementDirection,
 }
 
 impl HsvMovement {
-    pub(crate) fn is_enabled(self) -> bool {
+    pub fn is_enabled(self) -> bool {
         self.enabled
     }
 
-    pub(crate) fn speed_degrees_per_second(self) -> f32 {
+    pub fn speed_degrees_per_second(self) -> f32 {
         self.speed_degrees_per_second
     }
 
-    pub(crate) fn direction(self) -> HsvMovementDirection {
+    pub fn direction(self) -> HsvMovementDirection {
         self.direction
     }
 
-    pub(crate) fn stop(&mut self) {
+    pub fn stop(&mut self) {
         self.enabled = false;
     }
 
-    pub(crate) fn start(&mut self) {
+    pub fn start(&mut self) {
         self.enabled = true;
     }
 
-    pub(crate) fn set_speed(&mut self, speed: f32) {
+    pub fn set_speed(&mut self, speed: f32) {
         self.speed_degrees_per_second = speed.clamp(
             HSV_MOVEMENT_MIN_SPEED_DEGREES_PER_SECOND,
             HSV_MOVEMENT_MAX_SPEED_DEGREES_PER_SECOND,
         );
     }
 
-    pub(crate) fn set_direction(&mut self, direction: HsvMovementDirection) {
+    pub fn set_direction(&mut self, direction: HsvMovementDirection) {
         self.direction = direction;
     }
 
-    pub(crate) fn is_active(self, line_color: &LineColorConfig) -> bool {
+    pub fn is_active(self, line_color: &LineColorConfig) -> bool {
         self.enabled && matches!(line_color, LineColorConfig::HueCycle { .. })
     }
 }
@@ -77,7 +89,7 @@ impl Default for HsvMovement {
 
 /// Pure phase-advance computation. Kept as a free function so it remains testable
 /// without constructing signal infrastructure.
-pub(crate) fn advance_hsv_phase_degrees(
+pub fn advance_hsv_phase_degrees(
     phase_degrees: f32,
     speed_degrees_per_second: f32,
     dt_seconds: f32,
@@ -88,12 +100,11 @@ pub(crate) fn advance_hsv_phase_degrees(
 
 #[cfg(test)]
 mod tests {
-    use lsystem_core::LineColorConfig;
-
     use super::{
         HSV_MOVEMENT_MAX_SPEED_DEGREES_PER_SECOND, HSV_MOVEMENT_MIN_SPEED_DEGREES_PER_SECOND,
-        HsvMovement,
+        HsvMovement, HsvMovementDirection, advance_hsv_phase_degrees,
     };
+    use lsystem_core::LineColorConfig;
 
     #[test]
     fn set_speed_clamps_to_valid_range() {
@@ -119,5 +130,21 @@ mod tests {
         m.start();
         assert!(m.is_active(&LineColorConfig::DEFAULT_HUE_CYCLE));
         assert!(!m.is_active(&LineColorConfig::DEFAULT_SOLID));
+    }
+
+    #[test]
+    fn advance_wraps_forward() {
+        assert_eq!(
+            advance_hsv_phase_degrees(350.0, 20.0, 1.0, HsvMovementDirection::Forward),
+            10.0
+        );
+    }
+
+    #[test]
+    fn advance_wraps_reverse() {
+        assert_eq!(
+            advance_hsv_phase_degrees(10.0, 20.0, 1.0, HsvMovementDirection::Reverse),
+            350.0
+        );
     }
 }
